+++ nbproject/private/private.xml
@@ -4,9 +4,8 @@
     <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
         <group>
             <file>file:/Users/ben/Documents/420Project/src/celestialsoupuml/CelestialSoupUML.java</file>
-            <file>file:/Users/ben/Documents/420Project/src/celestialsoupuml/Boxes.java</file>
-            <file>file:/Users/ben/Documents/420Project/src/celestialsoupuml/Line.java</file>
             <file>file:/Users/ben/Documents/420Project/src/celestialsoupuml/ShapeEnum.java</file>
+            <file>file:/Users/ben/Documents/420Project/src/celestialsoupuml/WindowSingleton.java</file>
             <file>file:/Users/ben/Documents/420Project/src/celestialsoupuml/ShapeContainer.java</file>
         </group>
     </open-files>
