--- src/celestialsoupuml/Line.java
@@ -1,27 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package celestialsoupuml;
-
-/**
- *
- * @author bwjablon
- */
-public class Line {
-    
-    public int pointStartX;
-    public int pointStartY;
-    public int pointEndX;
-    public int pointEndY;
-    
-    public Line(int startX, int startY, int endX, int endY)
-    {
-        pointStartX = startX;
-        pointStartY = startY;
-        pointEndX = endX;
-        pointEndY = endY;
-    }
-    
-}
